feat: tutorial 4 solution

+++ 21cineplux/src/main/java/apap/tutorial/cineplux/service/BioskopServiceImpl.java
@@ -1,11 +1,15 @@
 package apap.tutorial.cineplux.service;
 
 import apap.tutorial.cineplux.model.BioskopModel;
+import apap.tutorial.cineplux.model.FilmModel;
 import apap.tutorial.cineplux.repository.BioskopDB;
+import apap.tutorial.cineplux.repository.FilmDB;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
+import java.time.LocalTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -16,8 +20,20 @@ public class BioskopServiceImpl implements BioskopService {
     @Autowired
     BioskopDB bioskopDB;
 
+    @Autowired
+    FilmDB filmDB;
+
     @Override
     public void addBioskop(BioskopModel bioskop) {
+
+        for (FilmModel film :
+                bioskop.getListFilm()) {
+            FilmModel filmModel = filmDB.getById(film.getNoFilm());
+            if (filmModel.getListBioskop() == null) {
+                filmModel.setListBioskop(new ArrayList<>());
+            }
+            filmModel.getListBioskop().add(bioskop);
+        }
         bioskopDB.save(bioskop);
     }
 
@@ -28,7 +44,7 @@ public class BioskopServiceImpl implements BioskopService {
 
     @Override
     public List<BioskopModel> getBioskopList() {
-        return bioskopDB.findAllByOrderByNamaBioskop();
+        return bioskopDB.findAll();
     }
 
     @Override
@@ -41,7 +57,20 @@ public class BioskopServiceImpl implements BioskopService {
     }
 
     @Override
-    public void removeBioskop(BioskopModel bioskop) {
-        bioskopDB.delete(bioskop);
+    public void deleteBioskop(BioskopModel bioskop) {
+        if (checkError(bioskop) == null) {
+            bioskopDB.delete(bioskop);
+        }
+    }
+
+    @Override
+    public String checkError(BioskopModel bioskop) {
+        if (LocalTime.now().isAfter(bioskop.getWaktuBuka()) && LocalTime.now().isBefore(bioskop.getWaktuTutup())) {
+            return "Bioskop sedang buka";
+        }
+        if (bioskop.getListPenjaga().size() > 0) {
+            return "Bioskop memiliki penjaga";
+        }
+        return null;
     }
 }
